+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Users can now add Logging and Monitoring agents to VPC Clusters
 - LogDNA and Sysdig platform logs and metrics are now enabled in the Power VS PoC templates
 
 ## 1.16.2